+++ example/lib/src/screens/features/features.dart
@@ -1,7 +1,8 @@
 /// Export all feature example screens.
 library;
 
-export 'chat_input_example.dart';
 export 'app_bar_example.dart';
+export 'chat_input_example.dart';
+export 'custom_builders_example.dart';
 export 'reply_example.dart';
 export 'search_example.dart';